+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.6.0] - 06-06-2022
+### Added
+- Add ability to clear Fastly cache via console command(`bin/magento core:clear-fastly`).
 ### Removed
 - Removed SVG supporting, move to separate module https://bitbucket.org/overdosedigital/module-magento-svg-support